small change
fixed time bug, changed int to floats

+++ src/Script.h
@@ -17,8 +17,8 @@ class Script
 {
 protected:
 	short scriptType;
-	int scriptCompareValue;
-	int scriptCurrentValue;
+	float scriptCompareValue;
+	float scriptCurrentValue;
 	std::string scriptStorage[7];
 	bool isScriptComplete;
 	Handlers* myHandles;
